PostLikeRepository 에 대한 테스트 코드 작성

+++ fastcampus/high-volume-data&traffic/fastcampus-mysql-init-project/src/main/java/com/example/fastcampusmysql/shcema/ddl.sql
@@ -64,8 +64,8 @@ create table Timeline
 create table PostLike
 (
     id        int auto_increment,
-    memberId  int      not null,
-    postId    int      not null,
+    memberId  bigint      not null,
+    postId    bigint      not null,
     createdAt datetime not null,
     constraint PostLike_id_uindex
         primary key (id)
